add rebar config
This patch add a way to build erlang_czmq using rebar. It only works
with rebar >= 2.0.0 .

The compilation environnement is created dynamically using the
rebar.config.scrip escript. if the environnment variable
ENABLE_CONFIG is set to no, it will try to build erlang_czmq using the
libraries installed on the system. It other cases it will fetch the
dependencies and build erlang_czmq statically like the Makefile does.